refactor(Car): 불필요하게 double로 선언된 변수들을 int로 변경함

+++ src/Car.java
@@ -43,12 +43,12 @@ public abstract class Car {
         return distance * totalMovement;
     }
 
-    public double totalFuel(int totalDistance) {
-        return (double) totalDistance / this.fuelEfficiency;
+    public int totalFuel(int totalDistance) {
+        return totalDistance / this.fuelEfficiency;
     }
 
-    public int countRefuel(double totalFuel) {
-        return (int) Math.ceil(totalFuel / this.fuelCapacity);
+    public int countRefuel(int totalFuel) {
+        return (int) Math.ceil((double) totalFuel / this.fuelCapacity);
     }
 
     public int totalCost(double totalFuel) {
